usePrevNextButtons moved into its own hook
