+++ src/hooks/usePrevNextButtons.js
@@ -1,0 +1,37 @@
+import { useCallback, useEffect, useState } from "react";
+
+export const usePrevNextButtons = (emblaApi, onButtonClick) => {
+    const [prevBtnDisabled, setPrevBtnDisabled] = useState(true);
+    const [nextBtnDisabled, setNextBtnDisabled] = useState(true);
+
+    const onPrevButtonClick = useCallback(() => {
+        if (!emblaApi) return;
+        emblaApi.scrollPrev();
+        if (onButtonClick) onButtonClick(emblaApi);
+    }, [emblaApi, onButtonClick]);
+
+    const onNextButtonClick = useCallback(() => {
+        if (!emblaApi) return;
+        emblaApi.scrollNext();
+        if (onButtonClick) onButtonClick(emblaApi);
+    }, [emblaApi, onButtonClick]);
+
+    const onSelect = useCallback((emblaApi) => {
+        setPrevBtnDisabled(!emblaApi.canScrollPrev());
+        setNextBtnDisabled(!emblaApi.canScrollNext());
+    }, []);
+
+    useEffect(() => {
+        if (!emblaApi) return;
+
+        onSelect(emblaApi);
+        emblaApi.on("reInit", onSelect).on("select", onSelect);
+    }, [emblaApi, onSelect]);
+
+    return {
+        prevBtnDisabled,
+        nextBtnDisabled,
+        onPrevButtonClick,
+        onNextButtonClick,
+    };
+};
